NetworkBrowser renamed and moved to a new directory:
