+++ src/components/Workspace/NetworkBrowserPanel/index.tsx
@@ -1,0 +1,1 @@
+export { NetworkBrowserPanel } from './NetworkBrowserPanel'
