+++ src/components/Workspace/SidePanel/PanelContents.tsx
@@ -10,11 +10,11 @@ import { TabPanel } from './TabPanel'
 
 export const getTabContents = (value: number): JSX.Element[] => {
   return [
-    <TabPanel label="test1" key={0} index={0} value={value}>
+    <TabPanel label="Hierarchy Viewer" key={0} index={0} value={value}>
       <ViewerPanel />
     </TabPanel>,
     <TabPanel label="test2" key={1} index={1} value={value}>
-      <Typography>test2</Typography>
+      <Typography variant="h5">Other new features here...</Typography>
     </TabPanel>,
   ]
 }
